fix: 내가 참가했던 대회
